Add Socket::Plain variant
Constant values are the same as desktop uses.

Invalid flags are converted to STARTTLS.

+++ src/constants.rs
@@ -213,15 +213,15 @@ pub const DC_LP_IMAP_SOCKET_PLAIN: i32 = 0x400;
 
 /// Connect to SMTP via STARTTLS.
 /// If this flag is set, automatic configuration is skipped.
-pub const DC_LP_SMTP_SOCKET_STARTTLS: usize = 0x10000;
+pub const DC_LP_SMTP_SOCKET_STARTTLS: i32 = 0x10000;
 
 /// Connect to SMTP via SSL.
 /// If this flag is set, automatic configuration is skipped.
-pub const DC_LP_SMTP_SOCKET_SSL: usize = 0x20000;
+pub const DC_LP_SMTP_SOCKET_SSL: i32 = 0x20000;
 
 /// Connect to SMTP unencrypted, this should not be used.
 /// If this flag is set, automatic configuration is skipped.
-pub const DC_LP_SMTP_SOCKET_PLAIN: usize = 0x40000;
+pub const DC_LP_SMTP_SOCKET_PLAIN: i32 = 0x40000;
 
 /// if none of these flags are set, the default is chosen
 pub const DC_LP_AUTH_FLAGS: i32 = DC_LP_AUTH_OAUTH2 | DC_LP_AUTH_NORMAL;
@@ -229,7 +229,7 @@ pub const DC_LP_AUTH_FLAGS: i32 = DC_LP_AUTH_OAUTH2 | DC_LP_AUTH_NORMAL;
 pub const DC_LP_IMAP_SOCKET_FLAGS: i32 =
     DC_LP_IMAP_SOCKET_STARTTLS | DC_LP_IMAP_SOCKET_SSL | DC_LP_IMAP_SOCKET_PLAIN;
 /// if none of these flags are set, the default is chosen
-pub const DC_LP_SMTP_SOCKET_FLAGS: usize =
+pub const DC_LP_SMTP_SOCKET_FLAGS: i32 =
     DC_LP_SMTP_SOCKET_STARTTLS | DC_LP_SMTP_SOCKET_SSL | DC_LP_SMTP_SOCKET_PLAIN;
 
 // QR code scanning (view from Bob, the joiner)
